создание таблицы взаимоотношений тренеров и клиентов

+++ src/main/kotlin/history_weight/HistoryWeightRouting.kt
@@ -12,17 +12,17 @@ import ru.alexbur.backend.base.success.toSuccess
 import ru.alexbur.backend.base.utils.DEFAULT_LIMIT
 import ru.alexbur.backend.base.utils.DEFAULT_OFFSET
 import ru.alexbur.backend.base.utils.getUserId
-import ru.alexbur.backend.client_card.service.ClientsCardService
 import ru.alexbur.backend.history_weight.mappers.HistoryWeightMapper
 import ru.alexbur.backend.history_weight.models.request.HistoryWeightUpdateRequest
 import ru.alexbur.backend.history_weight.models.response.HistoryWeightResponse
 import ru.alexbur.backend.history_weight.models.response.HistoryWeightsResponse
 import ru.alexbur.backend.history_weight.service.HistoryWeight
 import ru.alexbur.backend.history_weight.service.HistoryWeightService
+import ru.alexbur.backend.relationships.service.RelationshipsService
 
-fun Application.configureHistoryWeightRouting(
+internal fun Application.configureHistoryWeightRouting(
     service: HistoryWeightService,
-    clientsCardService: ClientsCardService,
+    relationshipsService: RelationshipsService,
     mapper: HistoryWeightMapper,
 ) {
 
@@ -31,7 +31,11 @@ fun Application.configureHistoryWeightRouting(
             put("/history-weight/{id}") {
                 val userId = call.getUserId() ?: return@put
                 val id = call.parameters["id"]?.toLong()
-                val historyWeight = getHistoryWeight(id, service, clientsCardService, userId) ?: return@put
+                val historyWeight = getHistoryWeight(id, service) ?: return@put
+                if (!relationshipsService.hasRelationship(coachId = userId, clientId = historyWeight.clientId)) {
+                    call.respond(HttpStatusCode.Forbidden, createBadRequestError(FitnessManagerErrors.EDITING_IS_PROHIBITED))
+                    return@put
+                }
                 val request = call.receive<HistoryWeightUpdateRequest>()
                 val isUpdated = service.update(historyWeight.id, request.weightGm)
                 if (!isUpdated) {
@@ -49,7 +53,11 @@ fun Application.configureHistoryWeightRouting(
             delete("/history-weight/{id}") {
                 val userId = call.getUserId() ?: return@delete
                 val id = call.parameters["id"]?.toLong()
-                val historyWeight = getHistoryWeight(id, service, clientsCardService, userId) ?: return@delete
+                val historyWeight = getHistoryWeight(id, service) ?: return@delete
+                if (!relationshipsService.hasRelationship(coachId = userId, clientId = historyWeight.clientId)) {
+                    call.respond(HttpStatusCode.Forbidden, createBadRequestError(FitnessManagerErrors.EDITING_IS_PROHIBITED))
+                    return@delete
+                }
                 val isDeleted = service.delete(historyWeight.id)
                 if (!isDeleted) {
                     call.respond(
@@ -61,22 +69,21 @@ fun Application.configureHistoryWeightRouting(
                 call.respond(HttpStatusCode.OK)
             }
 
-            get("/history-weight/{clientCardId}") {
+            get("/history-weight/{clientId}") {
                 val userId = call.getUserId() ?: return@get
-                val id = call.parameters["clientCardId"]?.toLong()
+                val id = call.parameters["clientId"]?.toLong()
                 if (id == null) {
                     call.respond(HttpStatusCode.BadRequest, createBadRequestError(FitnessManagerErrors.UNKNOWN_ID))
                     return@get
                 }
-                val clientCard = clientsCardService.getById(id, userId)
-                if (clientCard == null) {
-                    call.respond(HttpStatusCode.BadRequest, createBadRequestError(FitnessManagerErrors.UNKNOWN_CLIENT_CARD))
+                if (!relationshipsService.hasRelationship(coachId = userId, clientId = id)) {
+                    call.respond(HttpStatusCode.Forbidden, createBadRequestError(FitnessManagerErrors.EDITING_IS_PROHIBITED))
                     return@get
                 }
                 val limit = call.request.queryParameters["limit"]?.toIntOrNull() ?: DEFAULT_LIMIT
                 val offset = call.request.queryParameters["offset"]?.toIntOrNull() ?: DEFAULT_OFFSET
 
-                val data = service.getWeightByClientCardId(id, limit, offset)
+                val data = service.getWeightByClientId(id, limit, offset)
                 val response = mapper.map(data).toSuccess(HistoryWeightsResponse.serializer())
                 call.respond(HttpStatusCode.OK, response)
             }
@@ -87,8 +94,6 @@ fun Application.configureHistoryWeightRouting(
 private suspend fun RoutingContext.getHistoryWeight(
     id: Long?,
     service: HistoryWeightService,
-    clientsCardService: ClientsCardService,
-    userId: Long,
 ): HistoryWeight? {
     if (id == null) {
         call.respond(HttpStatusCode.BadRequest, createBadRequestError(FitnessManagerErrors.UNKNOWN_ID))
@@ -99,14 +104,6 @@ private suspend fun RoutingContext.getHistoryWeight(
         call.respond(
             HttpStatusCode.BadRequest,
             createBadRequestError(FitnessManagerErrors.UNKNOWN_HISTORY_WEIGHT)
-        )
-        return null
-    }
-    val clientCard = clientsCardService.getById(historyWeight.clientCardId, userId)
-    if (clientCard == null) {
-        call.respond(
-            HttpStatusCode.BadRequest,
-            createBadRequestError(FitnessManagerErrors.EDITING_IS_PROHIBITED)
         )
         return null
     }
